@@ -17,54 +17,54 @@
     "format": "prettier --write '{src,story}/**/*.{tsx,ts,jsx,js}'"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-feather": "^2.0.9",
-    "react-redux": "^7.2.6",
+    "react-redux": "^7.2.8",
     "redux": "^4.1.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@mdx-js/react": "^1.6.22",
-    "@storybook/addon-actions": "^6.3.9",
-    "@storybook/addon-docs": "^6.3.9",
-    "@storybook/addon-essentials": "^6.3.9",
-    "@storybook/addon-links": "^6.3.9",
+    "@babel/core": "^7.17.9",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@mdx-js/react": "^2.1.1",
+    "@storybook/addon-actions": "^6.4.21",
+    "@storybook/addon-docs": "^6.4.21",
+    "@storybook/addon-essentials": "^6.4.21",
+    "@storybook/addon-links": "^6.4.21",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/react": "^6.3.9",
-    "@testing-library/react": "^12.0.0",
-    "@types/jest": "^27.0.3",
+    "@storybook/react": "^6.4.21",
+    "@testing-library/react": "^13.0.0",
+    "@types/jest": "^27.4.1",
     "@types/node": "^16.11.9",
-    "@types/react": "^17.0.35",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/react-test-renderer": "^17.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.10.1",
-    "@typescript-eslint/parser": "^5.10.1",
-    "@vitejs/plugin-react": "^1.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "@vitejs/plugin-react": "^1.3.0",
     "@yarnpkg/plugin-compat": "^3.0.0",
     "@yarnpkg/pnpify": "^3.0.1",
     "autoprefixer": "^10.4.4",
     "babel-jest": "^27.5.1",
     "babel-loader": "^8.2.4",
-    "eslint": "^8.12.0",
+    "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-jest": "^26.1.3",
+    "eslint-plugin-jest": "^26.1.4",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-watch": "^8.0.0",
     "jest": "^27.5.1",
     "lint-staged": "^12.3.7",
     "postcss": "^8.4.12",
-    "prettier": "^2.6.1",
-    "react-test-renderer": "^17.0.2",
+    "prettier": "^2.6.2",
+    "react-test-renderer": "^18.0.0",
     "storybook-dark-mode": "^1.0.9",
     "tailwindcss": "^3.0.23",
     "ts-jest": "^27.1.4",
     "typescript": "^4.6.3",
-    "vite": "^2.8.6",
-    "webpack": "^5.70.0"
+    "vite": "^2.9.1",
+    "webpack": "^5.72.0"
   }
 }
